// src/api/models/category.ts
/**
 * Generated by orval v8.0.0-rc.0 🍺
 * Do not edit manually.
 * Backend Template API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryThumbnail } from './categoryThumbnail';

export interface Category {
  id?: string;
  name?: string;
  slug?: string;
  /** @nullable */
  url?: string | null;
  /**
   * @minimum 0
   * @nullable
   */
  sort_order?: number | null;
  /** @nullable */
  thumbnail_id?: string | null;
  /** @nullable */
  thumbnail?: CategoryThumbnail;
  created_at?: string;
  /** @nullable */
  created_by?: string | null;
  updated_at?: string;
  /** @nullable */
  updated_by?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  parent_id?: string | null;
}